Rename the drafts app to datasets
Will require a rebuild of the database, but should be a lot clearer what
is in the app now.
