+++ src/datasets/templates/datasets/edit_country.html
@@ -8,7 +8,7 @@
 {% block inner_content %}
 {% with countries=form.countries %}
 
-  {% include 'drafts/includes/error_box.html' %}
+  {% include 'datasets/includes/error_box.html' %}
 
   <h1 class="heading-large">
     Choose an area
