+++ src/datasets/templates/datasets/edit_frequency.html
@@ -7,7 +7,7 @@
 
 {% block inner_content %}
 
-  {% include 'drafts/includes/error_box.html' %}
+  {% include 'datasets/includes/error_box.html' %}
 
 
   <h1 class="heading-large">
